Add NameSpace And Code Convention
네임스페이스 추가 및 코드 컨벤션

+++ Assets/Scripts/Controller/BossController.cs
@@ -18,12 +18,12 @@ namespace Controller
             animSpeed = 0f;
             float startAnimSpeed =animinfo.StartAnimationSpeed;
             animSpeed = Mathf.Lerp(startAnimSpeed, 0f, (float)(elapsedTime / (animinfo.AnimLength * animinfo.DecelerationRatio)));
-            Anim.speed = (float)animSpeed;
+            Anim.speed = animSpeed;
             bool finished = animSpeed <= animinfo.AnimStopThreshold&& isCheckattackIndicatorFinish == true;
             if (finished)
             {
                 animSpeed = startAnimSpeed;
-                Anim.speed = (float)animSpeed;
+                Anim.speed = animSpeed;
             }
             return finished;
         }
